feat: rename auth route file and update references; enhance  authentication tests
