--- backend/src/routes/auth.js
@@ -1,7 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const authController = require('../controllers/authController');
-
-router.post('/google', authController.googleAuth);
-
-module.exports = router;
