+++ backend/src/tests/authRoutes.test.js
@@ -1,6 +1,6 @@
 const request = require('supertest');
 const express = require('express');
-const authRoutes = require('../routes/auth');
+const authRoutes = require('../routes/authRoutes');
 const authController = require('../controllers/authController');
 
 // Mock the auth controller
